MyPy Fixes and Enable CI Checks (#97)
This PR addresses all of the MyPy errors that we currently have. I've adjusted the mypy settings to match that of the django-stubs project https://github.com/typeddjango/django-stubs/blob/master/mypy.ini. For now I've delete all of our generated stubs as we didn't further refine them once generated. I've left the folder though in case we do want to be more prescriptive of untyped libs in the future.

Monday Tasks:
    CI - Set up Python MyPy Check
    CI - Set up Python Testing
    CI - Set up Python Linting
    Resolve MyPy type error so we can block on CI