Add Apollo Client to FE (#118)
This pull request integrates the Apollo Client into our frontend for efficient data management using GraphQL. It includes the base client for handling GraphQL queries and incorporates Apollo Link Rest for managing REST API calls. This setup enables the use of a unified client for both GraphQL and REST requests.

Additionally, we have implemented a csrfLink to manage the CSRF token for requests. The storage backend has undergone a refactoring process to facilitate its use with our custom csrfLink.

Comprehensive tests have been added to verify the proper functionality of the csrfLink.

This PR is a subset of the changes originally proposed in PR #114 from the BetterAngelsLA monorepo. It focuses specifically on the Apollo Client integration to streamline the review process. Subsequent updates in PR #114 will address the creation of queries, mutations, and the migration of existing REST calls to the Apollo Client.