Remove ADMIN_RESUMABLE_CHUNK_STORAGE, use default
This was a very stupid error on my part: I had left the custom storage
for the in-flight upload chunks as In-memory, which means any file
larger than one chunk would never work. I've changed it to use the
default_storage for the field in question, which should be correct.