Async upload (#710)
* Add async file upload to Attachment model

This is using an unsupported package that I manually updated to work
with Django 5. However, the entire package is small enough that it is
fairly easy to parse and understand, and thus modify and fix.

* Remove ADMIN_RESUMABLE_CHUNK_STORAGE, use default

This was a very stupid error on my part: I had left the custom storage
for the in-flight upload chunks as In-memory, which means any file
larger than one chunk would never work. I've changed it to use the
default_storage for the field in question, which should be correct.

* Add reference to python-magic

This is a library that is used to identify the MIME type of uploaded
files.

---------

Co-authored-by: Paul Vecchio <paul@betterangels.la>